made code stricter

+++ kernel/init/trap.c
@@ -41,8 +41,8 @@ void init_pit() {
 		// Set the PIT frequency to 256Hz
 		divisor = 1193180 / 256;
 		outb(0x43, 0x36);
-		outb(0x40, divisor & 0xFF);
-		outb(0x40, divisor >> 8);
+		outb(0x40, (uint8_t) (divisor & 0xFF));
+		outb(0x40, (uint8_t) (divisor >> 8));
 
 	cursek(74, -1);
 	printk("[done]");
